# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk mencari apotek terdekat, reservasi obat, dan transaksi pembayaran",
        "title": "Medifinder API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "cs.medifinder@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/apotek": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update informasi profil apotek termasuk jam operasional dan kontak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apotek"
                ],
                "summary": "Update Profil Apotek Sendiri",
                "parameters": [
                    {
                        "description": "Data update apotek",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateApotekRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: profil apotek berhasil diupdate",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: format data tidak valid / pesan error lainnya",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Membuat data apotek baru sekaligus upload foto profil",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apotek"
                ],
                "summary": "Tambah Apotek Baru",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama Apotek",
                        "name": "nama",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Alamat Lengkap",
                        "name": "alamat",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Latitude",
                        "name": "latitude",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Format HH:mm",
                        "name": "jam_buka",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Format HH:mm",
                        "name": "jam_tutup",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "File Foto (.jpg, .png)",
                        "name": "photo",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/foto": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Upload foto profil untuk apotek milik admin yang sedang login",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apotek"
                ],
                "summary": "Update Foto Apotek",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File Foto Apotek (.jpg, .png)",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/apotek": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data apotek yang dimiliki oleh admin yang sedang login",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Apotek"
                ],
                "summary": "Lihat apotek milik admin",
                "responses": {
                    "200": {
                        "description": "data apotek",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "apotek tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/apotek/nearby": {
            "get": {
                "description": "Mengambil daftar apotek berdasarkan lokasi user (latitude \u0026 longitude)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apotek"
                ],
                "summary": "Cari apotek terdekat",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitude user",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude user",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Radius pencarian dalam km (default 5, max 50)",
                        "name": "radius",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit data",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list apotek + pagination",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.UpdateApotekRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Jl. Merdeka No. 45"
                },
                "deskripsi": {
                    "type": "string",
                    "example": "Apotek buka setiap hari."
                },
                "jam_buka": {
                    "type": "string",
                    "example": "08:00:00"
                },
                "jam_tutup": {
                    "type": "string",
                    "example": "22:00:00"
                },
                "latitude": {
                    "type": "number",
                    "example": -6.2
                },
                "longitude": {
                    "type": "number",
                    "example": 106.816666
                },
                "nama": {
                    "type": "string",
                    "example": "Apotek Sehat Sejahtera"
                },
                "phone_number": {
                    "type": "string",
                    "example": "081234567890"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://example.com/photo.jpg"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}